t/xml/cdata-1.dat

#data
<![CDATA[]]>
#errors
1;1;cdata outside of root element;;;m
1;12;no root element;;;m
#document

#data
<![CDATA[  ]]>
#errors
1;1;cdata outside of root element;;;m
1;14;no root element;;;m
#document
| "  "

#data
<![CDATA[ab]]>
#errors
1;1;cdata outside of root element;;;m
1;14;no root element;;;m
#document
| "ab"

#data
a<![CDATA[bb]]>ccc
#errors
1;1;text outside of root element;;;m
1;18;no root element;;;m
#document
| "abbccc"

#data
<a/><![CDATA[]]>
#errors
1;5;cdata outside of root element;;;m
#document
| <{} a>

#data
<a/><![CDATA[  ]]>
#errors
1;5;cdata outside of root element;;;m
#document
| <{} a>
| "  "

#data
<a/><![CDATA[aa]]>
#errors
1;5;cdata outside of root element;;;m
#document
| <{} a>
| "aa"

#data
<a/><![CDATA[]]>  a
#errors
1;5;cdata outside of root element;;;m
#document
| <{} a>
| "  a"

#data
<![CDATA[
#errors
1;1;cdata outside of root element;;;m
1;9;no mse;;;m
1;9;no root element;;;m
#document

#data
<![CDATA[a
#errors
1;10;no mse;;;m
1;1;cdata outside of root element;;;m
1;10;no root element;;;m
#document
| "a"

#data
<![CDATA[a]
#errors
1;11;no mse;;;m
1;1;cdata outside of root element;;;m
1;11;no root element;;;m
#document
| "a]"

#data
<![CDATA[a]]
#errors
1;12;no mse;;;m
1;1;cdata outside of root element;;;m
1;12;no root element;;;m
#document
| "a]]"

#data
<![CDATA[a]]><![CDATA[b]]><a/>
#errors
1;1;cdata outside of root element;;;m
#document
| "ab"
| <{} a>

#data escaped
<![CDATA[\u0000]]>
#errors
1;1;cdata outside of root element;;;m
1;1;NULL;;;m
1;13;no root element;;;m
#document escaped
| "\uFFFD"

#data escaped
<!DOCTYPE p><![CDATA[\u0000]]><p></p>
#errors
1;13;cdata outside of root element;;;m
1;13;NULL;;;m
#document escaped
| <!DOCTYPE p>
| "\uFFFD"
| <{} p>

#data escaped
<p><![CDATA[\u0000]]></p>
#errors
1;4;NULL;;;m
#document escaped
| <{} p>
|   "\uFFFD"

#data escaped
<p></p><![CDATA[\u0000]]>
#errors
1;8;NULL;;;m
1;8;cdata outside of root element;;;m
#document escaped
| <{} p>
| "\uFFFD"

#data escaped
<p></p><![CDATA[a\u0000b]]>
#errors
1;8;NULL;;;m
1;8;cdata outside of root element;;;m
#document escaped
| <{} p>
| "a\uFFFDb"
